Refactor regenerating resources.
Create an abstract class RegeneratingResource that sets up the
regeneration portion and let Energy, Focus, and Mana inherit this class.

Let Mana carry fractions and tick every 2 seconds instead of 5.

Let Character::gain_x add player reaction events instead of Melee
Attacks, ResourceGain.

+++ Character/Character.cpp
@@ -479,9 +479,9 @@ void Character::lose_focus(const unsigned)
 
 }
 
-unsigned Character::get_mp5_from_spirit() const {
+double Character::get_mp5_from_spirit() const {
     check(false, "Not implemented");
-    return 0;
+    return 0.0;
 }
 
 void Character::reset_resource() {
